Padronizado projeto para ser trabalhado
