--- src/App.css
@@ -1,24 +1,0 @@
-<<<<<<< HEAD
-=======
-*{
-  margin: 0;
-  padding: 0;
-}
-
-.produtos {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr 1fr 1fr 1fr;
-  justify-content: center;
-  align-items: center;
-  border: 1px solid black;
-  max-width: 50px;
-  max-height: 30px;
-
-}
-
-.imagem {
-  max-width: 20px;
-  max-height: 20px;
-}
-
->>>>>>> testes-antes-da-master
