+++ src/components/styles/App.css
@@ -1,0 +1,54 @@
+.app-container {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-evenly;
+  align-items: center;
+  width: 100vh;
+  height: 100vh;
+  box-sizing: border-box;
+  border: 1px solid black;
+}
+
+.filtros-container {
+  display: flex;
+  box-sizing: border-box;
+  flex-direction: column;
+  align-items: center;
+  border: 1px solid black;
+  width: 200px;
+  height: 100vh;
+  padding: 8px;
+}
+
+.filtros-container > label {
+  margin-top: 4px;
+}
+
+.produtos-container {
+  display: grid;
+  grid-template-columns: 1fr 3fr 1fr;
+  padding: 16px;
+  gap: 8px;
+}
+
+.card-produtos {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 16px;
+  padding: 16px;
+}
+
+.produtos-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 0 16px;
+}
+
+.produtos-grid {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 16px;
+  padding: 16px;
+}
